requirements.txt
# 智能内容情报 Agent 系统 - 依赖清单

# 核心爬虫
twikit>=2.0.0

# 数据处理
pandas>=2.0.0
openpyxl>=3.1.0

# 网页解析
beautifulsoup4>=4.12.0
requests>=2.31.0

# 模糊匹配
thefuzz>=0.20.0
python-Levenshtein>=0.23.0  # 加速 thefuzz

# 环境变量管理 (可选)
python-dotenv>=1.0.0
